+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added basic description text
+- Added keys and interactivity to well construction table, moved next to  drawing.
 
 ## [0.1.0]
 ### Added